initial commit
pretty much moved everything, we just need add/delete/rename attribute
